requirements.txt

lightning-flash==0.5.0
Markdown==3.3.4
MarkupSafe==2.0.1
multidict==5.1.0
munch==2.5.0
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.2.5
Pillow==8.3.2
pretrainedmodels==0.7.4
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyDeprecate==0.3.2
pyparsing==2.4.7
pystiche==1.0.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytorch-lightning==1.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
segmentation-models-pytorch==0.2.0
six==1.16.0
starlette==0.14.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
timm==0.4.12
toolz==0.11.1
torch==1.9.0
torchmetrics==0.5.0
torchvision==0.10.0
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.13.4
uvloop==0.16.0
watchgod==0.7
websockets==8.1
Werkzeug==2.0.1
yarl==1.6.3
